backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.18.0
alembic>=1.14.0
pydantic>=1.10.12
PyJWT>=2.10.0
httpx>=0.28.0
python-multipart>=0.0.20
redis>=5.2.0
aiofiles>=24.1.0
playwright>=1.49.0
groq>=0.9.0
